worked on #8.
I tested this last night and it seems that at midnight UTC, it switches the day in the start date variable. This messes up all the queries since they all use it. So for a central time unit, the start date would be 2016-07-01 05:00:00 UTC at 7pm local. This means that no data would be shown until midnight local time